+++ com.unity.render-pipelines.lightweight/Editor/ShaderGraph/LightWeightPBRSubShader.cs
@@ -4,12 +4,13 @@ using System.IO;
 using System.Linq;
 using UnityEditor.Graphing;
 using UnityEditor.ShaderGraph;
-using UnityEngine.Experimental.Rendering.LightweightPipeline;
 using UnityEngine.Rendering;
-
-namespace UnityEditor.Experimental.Rendering.LightweightPipeline
+using UnityEngine.Rendering.LWRP;
+    
+namespace UnityEditor.Rendering.LWRP
 {
     [Serializable]
+    [FormerName("UnityEditor.Experimental.Rendering.LightweightPipeline.LightWeightPBRSubShader")]
     [FormerName("UnityEditor.ShaderGraph.LightWeightPBRSubShader")]
     class LightWeightPBRSubShader : IPBRSubShader
     {