+++ com.unity.render-pipelines.lightweight/Editor/UpgradeCommon.cs
@@ -1,4 +1,4 @@
-namespace UnityEditor.Experimental.Rendering.LightweightPipeline
+namespace UnityEditor.Rendering.LWRP
 {
     public enum UpgradeSurfaceType
     {