+++ com.unity.render-pipelines.lightweight/Runtime/ForwardRendererSetup.cs
@@ -1,10 +1,11 @@
 using System;
 using System.Collections.Generic;
 using UnityEngine.Rendering;
+using UnityEngine.Rendering.LWRP;
 
-namespace UnityEngine.Experimental.Rendering.LightweightPipeline
+namespace UnityEngine.Experimental.Rendering.LWRP
 {
-    internal class DefaultRendererSetup : IRendererSetup
+    internal class ForwardRendererSetup : IRendererSetup
     {
         private DepthOnlyPass m_DepthOnlyPass;
         private MainLightShadowCasterPass m_MainLightShadowCasterPass;