+++ com.unity.render-pipelines.lightweight/Runtime/Passes/DepthOnlyPass.cs
@@ -1,7 +1,8 @@
 using System;
 using UnityEngine.Rendering;
+using UnityEngine.Rendering.LWRP;
 
-namespace UnityEngine.Experimental.Rendering.LightweightPipeline
+namespace UnityEngine.Experimental.Rendering.LWRP
 {
     /// <summary>
     /// Render all objects that have a 'DepthOnly' pass into the given depth buffer.