+++ com.unity.testing.srp.lightweight/Tests/Scenes/040_UpgradeScene/Scripts/Editor/LWSceneConverter.cs
@@ -2,7 +2,6 @@
 using System.Collections.Generic;
 using UnityEngine;
 using UnityEngine.Rendering;
-using UnityEngine.Experimental.Rendering;
 using UnityEngine.SceneManagement;
 using UnityEditor;
 using UnityEditor.Experimental.Rendering.LightweightRP;