+++ com.unity.testing.srp.lightweight/Tests/Scenes/045_CustomLWPipe/CustomLWPipe.cs
@@ -4,31 +4,11 @@ using UnityEditor;
 using UnityEditor.ProjectWindowCallback;
 #endif
 using UnityEngine;
-using UnityEngine.Experimental.Rendering;
-using UnityEngine.Experimental.Rendering.LightweightPipeline;
+using UnityEngine.Experimental.Rendering.LWRP;
 using UnityEngine.Rendering;
-
-public class CustomLWPipe : LightweightRendererSetup
-{
-    #if UNITY_EDITOR 
-    [MenuItem("Assets/Create/Rendering/CustomLWSetup", priority = CoreUtils.assetCreateMenuPriority1)]
-    static void CreateCustomLwPipe()
-    {
-        ProjectWindowUtil.StartNameEditingIfProjectWindowExists(0, CreateInstance<CreateCustomLWPipe>(),
-            "CustomLWPipe.asset", null, null);
-    }
-    
-    [System.Diagnostics.CodeAnalysis.SuppressMessage("Microsoft.Performance", "CA1812")]
-    internal class CreateCustomLWPipe : EndNameEditAction
-    {
-        public override void Action(int instanceId, string pathName, string resourceFile)
-        {
-            var instance = CreateInstance<CustomLWPipe>();
-            AssetDatabase.CreateAsset(instance, pathName);
-        }
-    }
-    #endif
-    
+using UnityEngine.Rendering.LWRP;
+public class CustomLWPipe : MonoBehaviour, IRendererSetup
+{  
     private SetupForwardRenderingPass m_SetupForwardRenderingPass;
     private CreateLightweightRenderTexturesPass m_CreateLightweightRenderTexturesPass;
     private SetupLightweightConstanstPass m_SetupLightweightConstants;
@@ -50,7 +30,7 @@ public class CustomLWPipe : LightweightRendererSetup
         m_Initialized = true;
     }
 
-    public override void Setup(ScriptableRenderer renderer, ref RenderingData renderingData)
+    public void Setup(ScriptableRenderer renderer, ref RenderingData renderingData)
     {
         Init();